Fix unit tests and warnings
