+++ preferences/managers.py
@@ -7,20 +7,20 @@ class SingletonManager(models.Manager):
     """
     Returns only a single preferences object per site.
     """
-    def get_query_set(self):
+    def get_queryset(self):
         """
         Return the first preferences object for the current site.
         If preferences do not exist create it.
         """
-        queryset = super(SingletonManager, self).get_query_set()
+        queryset = super(SingletonManager, self).get_queryset()
 
         # Get current site
         current_site = None
-        if getattr(settings, 'SITE_ID', None) != None:
+        if getattr(settings, 'SITE_ID', None) is not None:
             current_site = Site.objects.get_current()
 
         # If site found limit queryset to site.
-        if current_site != None:
+        if current_site is not None:
             queryset = queryset.filter(sites=settings.SITE_ID)
 
         try:
@@ -28,7 +28,7 @@ class SingletonManager(models.Manager):
         except self.model.DoesNotExist:
             # Create object (for current site) if it doesn't exist.
             obj = self.model.objects.create()
-            if current_site != None:
+            if current_site is not None:
                 obj.sites.add(current_site)
 
         return queryset
