+++ preferences/tests/test_settings.py
@@ -1,0 +1,34 @@
+from django.core.wsgi import get_wsgi_application
+
+INSTALLED_APPS = [
+    'django.contrib.auth',
+    'django.contrib.sessions',
+    'django.contrib.contenttypes',
+    'django.contrib.admin',
+    'django.contrib.sites',
+    'preferences',
+    'preferences.tests',
+]
+
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+    }
+}
+
+MIDDLEWARE_CLASSES = (
+    'django.middleware.common.CommonMiddleware',
+    'django.middleware.csrf.CsrfViewMiddleware',
+    'django.contrib.sessions.middleware.SessionMiddleware',
+    'django.contrib.auth.middleware.AuthenticationMiddleware',
+)
+
+TEMPLATE_CONTEXT_PROCESSORS = (
+    'preferences.context_processors.preferences_cp',
+)
+
+ROOT_URLCONF = 'preferences.tests.urls'
+
+SECRET_KEY = "test_secret_key"
+
+application = get_wsgi_application()
