+++ setup.py
@@ -4,7 +4,7 @@ setup(
     name='django-preferences',
     version='0.0.6',
     description='Django app allowing users to set app specific preferences through the admin interface.',
-    long_description = open('README.rst', 'r').read() + open('AUTHORS.rst', 'r').read() + open('CHANGELOG.rst', 'r').read(),
+    long_description=open('README.rst', 'r').read() + open('AUTHORS.rst', 'r').read() + open('CHANGELOG.rst', 'r').read(),
     author='Praekelt Foundation',
     author_email='dev@praekelt.com',
     license='BSD',
@@ -15,7 +15,7 @@ setup(
         'django-setuptest>=0.0.6',
     ],
     include_package_data=True,
-    install_requires = [
+    install_requires=[
         'Django',
     ],
     classifiers=[
